rqspinlock: Do not destroy queue with AA deadlocks
Amery highlighted that the current behavior for draining the waiter
queue can be too conservative for rqspinlock, as any deadlock (AA or
ABBA) will lead to the queue being signalled and waiters being told to
exit.

In case a head waiter encounters an AA deadlock, but the rest of the
waiters are perfectly capable of obtaining the lock, they will still be
signalled and fail the lock acquisition. This is unnecessary, as an AA
deadlock at the head should be detected instantly, and does not require
the subsequent waiters in the queue to drain themselves to allow
everyone to make forward progress. In case an AA or ABBA condition
happen simultaneously, the next head if any can detect the ABBA
condition and trigger the queue drainage explicitly, but it needn't be
done when the head is only subjected to an AA condition.

Adapt the checks to report -EBUSY for AA deadlock, and convert that back
to -EDEADLK to keep the reported error the same for the caller, but
handle that differently for the head waiter. We can simply signal the
next node to become the next head, and not attempt acquiring the lock.
The next head will then repeat the spin waiting loop on locked and
pending bit if needed, and then attempt to set_locked() to gain lock
ownership.

Cc: Amery Hung <ameryhung@gmail.com>
Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>