rqspinlock: Switch to rqnode and res_mcs_spinlock
In preparation for changes to how queue destruction occurs on the slow
path, refactor rqspinlock to use a separate MCS lock struct and a
different qnode type, since we will introduce a new field for subsequent
changes.

In addition to this, create {encode,decode}_tail_rqnode accessors, that
allow encoding a special sentinel value in addition to the normal tail
and empty state (0). This is useful in subsequent changes as well, since
we'll need a dummy tail value in addition to normal and empty ones.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>