rqspinlock: Address possible edge case due to queue reset
The current queue reset strategy can cause potential confusion for the
tail if the head races ahead in resetting the queue before the sentinel
state is set. This can cause the tail into waiting aimlessly for the
next pointer to be populated. It is not easy to address this problem
without other significant changes to the recovery algorithm. Instead,
give up on resetting the queue.

If a new waiter notices the sentintel value, it just means that the
setting of the value was the severage of the old queue from any
potential future queueing waiters, and the waiter noticing -1:0 as
prev can consider itself as the head.

There is yet another problem here though, the head of the old queue
can now race with the new head for possible access to the lock, as
soon as both notice the locked bit to be dropped.

This requires some sort of handling, which is TBD (most likely in the
new waiter joining the queue).

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>