rqspinlock: Back propagate deadlock signals from successors
Allow successors that do not sample deadlocks and timeouts and signal
the rest of the waiter chain (in case of timeouts) to publish whether
they are locally seeing a deadlock or timeout. This is useful in the
cases where successors block the progress of the owner or pending
waiter, which in turn prevents progress of the head waiter and thus
the successor.

For the common case involving the same lock, such a case is prevented by
the AA check done on entry into the queueing path.

CPU 0
<NMI>
lock A -> non-head waiter
</NMI>
owner of A or pending waiter of A

CPU 1
lock A -> head waiter -> times out after 250ms if A doesn't exit

Right now, this lock A attempt on CPU 0 will simply fail preventing the
250 ms timeout. Now imagine a more contrived case involving two locks,
where typical ABBA checks in case of non-head NMI waiters would have
worked.

CPU 0
<NMI>
lock B -> non-head waiter
</NMI>
owner of A

CPU 1
<NMI>
lock A -> non-head waiter
</NMI>
owner of B

CPU 2
lock B -> head waiter

CPU 3
lock A -> head waiter

In such a case, we could try optimistic check_deadlock_ABBA() before
entry, but it's always going to remain racy, and doesn't preclude the
timeout in case the ABBA forms right after the check. The only proper
way to solve this is for the NMI non-head waiters to sample for ABBA
after they have queued themselves. This commit tries to address such
a scenario.

Each non-head waiter will sample for such deadlocks and store a
RES_HUNG_VAL in its MCS node's locked field. This can of course be
overwritten by the predecessor signalling 1 (become head) or 2
(timeout).

The predecessor then uses this field to understand whether the next node
has sampled a deadlock, and treats it as a deadlock for itself. To
expedite the recovery, the deadlock signal is upgraded to a -ETIMEDOUT,
otherwise waiter chain moves one waiter at a time until it reaches the
successor in the chain that is encountering a deadlock.

Once this is done, the signal eventually is back propagated to the head
that will encounter it as an -ETIMEDOUT, and terminate the whole waiting
queue, leading to a faster recovery than a proper timeout at the head
that would wait for the entire 250 ms.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>